+++ Changelog.md
@@ -80,6 +80,7 @@
 * Values of `DocxShapeSize#flip_horizontal` and 
 `DocxShapeSize#flip_vertical` are Boolean
 * Fix crash on gradient stop with unknown SchemeColor ([ooxml_parser#571](https://github.com/ONLYOFFICE/ooxml_parser/issues/571))
+* Fix `Presentation.with_data?` for shape with preset geometry ([ooxml_parser#573](https://github.com/ONLYOFFICE/ooxml_parser/issues/573))
 
 ## 0.4.1 (2018-03-01)
 ### Fixes